Beginning Circle Creation and Price Data
I created a shell of code that will help in the creation of circles with
the radius dependent on the price amount data. Still, there are some
errors that need to be fixed, where the functions I'm calling aren't
showing up (despite the fact that they work in the console).